requirements.txt
# WireGuard MikroTik VPS Project Dependencies
# Python packages required for the management system

# Core dependencies
requests>=2.28.0
psutil>=5.9.0

# Optional Django integration (if using with Django project)
# django>=3.2.0
# librouteros>=3.3.0

# Database Integration
supabase>=2.0.0
psycopg2-binary>=2.9.0

# Development and testing
pytest>=7.0.0
pytest-cov>=4.0.0

# System monitoring
colorama>=0.4.6
tabulate>=0.9.0

# Network utilities
netifaces>=0.11.0
